Move "nice-to-have" scripts to sandbox
This commit moves `conformal_map_taylor_coefficients.jl` and
`complex_jacobi_ellptic.jl` from main package to `sandbox`, allowing us to
remove most dependencies of `CubedSphere.jl`. `sandbox` now contains a Julia
environment with the packages needed to execute scripts in the folder.

Technically, this is a breaking change, so I bumped to 0.3.0.

No changes to the documentation seem to be required.