Bump to 0.3.2
https://github.com/JuliaDiff/TaylorSeries.jl/pull/345 will add support
to a newer version of IntervalArithmetics. This is needed to support
recent versions of Makie.

https://github.com/JuliaDiff/TaylorSeries.jl/pull/345 bumps the version
of TaylorSeries to 0.19, and this commit expands compat for that.

CubedSphere.jl does not use IntervalArithmetics, so it is still
compatible with 0.19.

I tested 345 locally and the test suite for CubedSphere passes.

(We can merge this after TaylorSeries 0.19 is released)